refactor formating
