+++ src/main/java/fitnesbot/exeptions/usererrors/UserAlreadyExistsError.java
@@ -1,4 +1,4 @@
-package fitnesbot.exeptions.UserErrors;
+package fitnesbot.exeptions.usererrors;
 
 public class UserAlreadyExistsError extends UserErrors {
     private final long userId;
